requirements.txt
gensim==3.8.3
Keras==2.1.2
matplotlib==3.1.0
numpy==1.15.4
pandas==1.0.4
scikit-learn==0.19.1
scipy==1.4.1
sklearn==0.0
tensorflow-gpu==1.12.0




